Validate Phase 0 schema against a real Supabase project; fix critical RPC hole
Adds the Supabase CLI as a pinned dev dependency (2.117.0) and validates the
Phase 0 schema end-to-end against a throwaway staging project (pawcue-staging,
ref ccqitncejkgsaborfubc) instead of trusting review-by-eye.

Doing that immediately found a critical, remotely exploitable bug that both
manual SQL review and Postgres-grammar syntax checking had passed:

  merge_guest_session() is SECURITY DEFINER and re-parents one account's dogs,
  subscriptions and entitlements onto another. PostgREST auto-exposes every
  public function at /rest/v1/rpc/<name>, and Postgres grants EXECUTE to PUBLIC
  by default -- which the migration never revoked. Anyone holding the anon key
  (shipped inside the mobile app by design, therefore public) could take
  ownership of an arbitrary victim's dogs, training history, subscription and
  entitlement -- premium included -- with no authentication at all.

  Confirmed by exploit, not inferred: as the anon role, a test victim's dog was
  successfully transferred to a test attacker. Fixed by revoking EXECUTE from
  public/anon/authenticated and adding in-function guards (no self-merge,
  source must be anonymous, target must be permanent). Authorization stays the
  Edge Function's job -- it must verify both the caller's authenticated JWT and
  the guest's anonymous-session JWT -- but the function is no longer reachable
  without the service role. Written up in SECURITY.md.

Other issues found by execution and fixed:
- seed.sql failed on a UNION type mismatch (untyped NULL resolves to text,
  colliding with an integer branch). db push never runs seed.sql, only db reset
  does, so this had never actually executed before.
- 8 foreign keys had no covering index, contradicting DATABASE.md's explicit
  claim. Unindexed FKs force a seq scan of the child table on every cascading
  DELETE, and account deletion cascades across most of this schema.
- All 28 RLS policies re-evaluated auth.uid() per row; now wrapped in a scalar
  subselect (Supabase's auth_rls_initplan lint).
- set_updated_at() had a mutable search_path.
- Domain/schema drift: profiles.is_anonymous was missing from userSchema even
  though merge_guest_session validates on it, and app_events.created_at (server
  receive time, distinct from client occurred_at) was missing from appEventSchema.
- The i18n tests did `await i18n.init` -- awaiting a method reference, not a
  promise, so they never waited for anything. Caught by type-aware lint.
- ESLint was referenced in scripts but never installed, so the repo's own gate
  could not have passed. Now installed and configured (flat config, type-aware).
- The gate script was named "ci", which pnpm v12 shadows with its built-in ci
  command; renamed to `pnpm verify`.

Added, and now part of the workflow:
- supabase/tests/rls_security.sql -- 49 checks that drop to the real anon and
  authenticated roles (asserting as a privileged role proves nothing, since the
  migration role bypasses RLS): public catalog access, own-data access,
  cross-user denial for read/update/delete/ownership-forgery, unauthorized
  writes to entitlements/subscriptions/purchase_events, the privilege-escalation
  regression, guest-merge correctness and idempotency, constraints, auth trigger,
  updated_at trigger, and account deletion (personal data hard-deleted;
  billing/audit rows retained with user_id nulled).
- Generated database types committed, guarded by schema-conformance.test.ts,
  which parses them and fails on drift in either direction.
- pnpm db:reset / db:test / db:advisors / db:verify / db:types.

Verification: migrations replay from zero and seed cleanly; 49/49 DB checks
pass; Supabase advisors report no issues (was 23 findings); `pnpm verify`
green (format, lint, typecheck, 77 unit tests).

No credentials committed -- DB password and project ref live only in a
gitignored .env.supabase.local.

+++ supabase/seed.sql
@@ -74,11 +74,13 @@ insert into lesson_steps (lesson_id, step_order, instruction_key, requires_click
   ('00000000-0000-4000-a000-0000000020a1', 2, 'lesson.sit.step3', true,  5);
 
 insert into lesson_steps (lesson_id, step_order, instruction_key, requires_clicker_press, repetition_target)
-select id, 0, slug || '.step1', false, null from lessons where slug not in ('name_game', 'sit')
+-- repetition_target is cast explicitly: an untyped NULL resolves to text and collides with the integer in the
+-- third branch ("UNION types text and integer cannot be matched").
+select id, 0, slug || '.step1', false, null::smallint from lessons where slug not in ('name_game', 'sit')
 union all
-select id, 1, slug || '.step2', true, null from lessons where slug not in ('name_game', 'sit')
+select id, 1, slug || '.step2', true, null::smallint from lessons where slug not in ('name_game', 'sit')
 union all
-select id, 2, slug || '.step3', false, 5 from lessons where slug not in ('name_game', 'sit');
+select id, 2, slug || '.step3', false, 5::smallint from lessons where slug not in ('name_game', 'sit');
 
 -- ---------------------------------------------------------------------------
 -- lesson_troubleshooting — the core differentiator (brief §7). Fully authored for Name Game and Sit;
